+++ CHANGES.md
@@ -4,6 +4,9 @@
 - Use new 3D and flood fill subset state classes from glue to make storing
   subsets much more efficient. [#301]
 
+- Work around an issue on certain graphics cards which causes volume
+  renderings to not appear correctly but instead of have stripe artifacts. [#303]
+
 0.9.2 (unreleased)
 ------------------